Add Proper Logging and Child Process Handler
Proper logging is added by configuring two root loggers. One is for
output purpose only, which lives in its own process. The other is
for passing all other log record to the output logger via a queue.
This set up is essential for recording the logging info in a local
file.